endedproject

+++ ArvinTav/Areas/User/Controllers/DashboardController.cs
@@ -48,21 +48,21 @@ namespace ArvinTav.Areas.User.Controllers
             DataLayer.User UpdateUser = new DataLayer.User();
             DataLayer.User user = userRepository.UserByPhoneNumber(User.Identity.Name);
 
-            if (fullRegsiterViewModel.Brand.Contains("'"))
+            if(string.IsNullOrEmpty(fullRegsiterViewModel.FullName) || string.IsNullOrEmpty(fullRegsiterViewModel.Brand) || string.IsNullOrEmpty(fullRegsiterViewModel.Email))
             {
-                ModelState.AddModelError("Brand", "نام تجاری نا معتبر است");
                 FullRegsiterViewModel fullReplyRegsiterViewModel = new FullRegsiterViewModel();
                 fullReplyRegsiterViewModel.FullName = user.FullName;
                 fullReplyRegsiterViewModel.Brand = user.Brand;
                 fullReplyRegsiterViewModel.Email = user.Email;
                 fullReplyRegsiterViewModel.Image = user.Image;
+                ViewBag.Erorr = "تمامی گزینه ها را پر کنید";
                 return View(fullRegsiterViewModel);
             }
             else
             {
-                if (fullRegsiterViewModel.Email.Contains("'"))
+                if (fullRegsiterViewModel.Brand.Contains("'"))
                 {
-                    ModelState.AddModelError("Email", "ایمیل نا معتبر است");
+                    ModelState.AddModelError("Brand", "نام تجاری نا معتبر است");
                     FullRegsiterViewModel fullReplyRegsiterViewModel = new FullRegsiterViewModel();
                     fullReplyRegsiterViewModel.FullName = user.FullName;
                     fullReplyRegsiterViewModel.Brand = user.Brand;
@@ -72,9 +72,9 @@ namespace ArvinTav.Areas.User.Controllers
                 }
                 else
                 {
-                    if (fullRegsiterViewModel.FullName.Contains("'"))
+                    if (fullRegsiterViewModel.Email.Contains("'"))
                     {
-                        ModelState.AddModelError("نام نا معتبر است", "تصویر با فرمت png و jpeg معتبر است");
+                        ModelState.AddModelError("Email", "ایمیل نا معتبر است");
                         FullRegsiterViewModel fullReplyRegsiterViewModel = new FullRegsiterViewModel();
                         fullReplyRegsiterViewModel.FullName = user.FullName;
                         fullReplyRegsiterViewModel.Brand = user.Brand;
@@ -84,27 +84,24 @@ namespace ArvinTav.Areas.User.Controllers
                     }
                     else
                     {
-                        if (ImageUp != null)
+                        if (fullRegsiterViewModel.FullName.Contains("'"))
                         {
-
-                            if (ImageUp.ContentType != "image/png" && ImageUp.ContentType != "image/jpeg" && ImageUp.ContentType != "image/jpg")
+                            ModelState.AddModelError("نام نا معتبر است", "تصویر با فرمت png و jpeg معتبر است");
+                            FullRegsiterViewModel fullReplyRegsiterViewModel = new FullRegsiterViewModel();
+                            fullReplyRegsiterViewModel.FullName = user.FullName;
+                            fullReplyRegsiterViewModel.Brand = user.Brand;
+                            fullReplyRegsiterViewModel.Email = user.Email;
+                            fullReplyRegsiterViewModel.Image = user.Image;
+                            return View(fullRegsiterViewModel);
+                        }
+                        else
+                        {
+                            if (ImageUp != null)
                             {
-                                ModelState.AddModelError("Image", "تصویر با فرمت png و jpeg معتبر است");
-                                FullRegsiterViewModel fullReplyRegsiterViewModel = new FullRegsiterViewModel();
-                                fullReplyRegsiterViewModel.FullName = user.FullName;
-                                fullReplyRegsiterViewModel.Brand = user.Brand;
-                                fullReplyRegsiterViewModel.Email = user.Email;
-                                fullReplyRegsiterViewModel.Image = user.Image;
-                                return View(fullRegsiterViewModel);
-                            }
-                            else
-                            {
 
-                                var img = Bitmap.FromStream(ImageUp.InputStream);
-
-                                if (img.Width > 600 || img.Height > 600)
+                                if (ImageUp.ContentType != "image/png" && ImageUp.ContentType != "image/jpeg" && ImageUp.ContentType != "image/jpg")
                                 {
-                                    ModelState.AddModelError("Image", "طول و عرض تصویر نباید بیشتر از 600 پیکسل باشد");
+                                    ModelState.AddModelError("Image", "تصویر با فرمت png و jpeg معتبر است");
                                     FullRegsiterViewModel fullReplyRegsiterViewModel = new FullRegsiterViewModel();
                                     fullReplyRegsiterViewModel.FullName = user.FullName;
                                     fullReplyRegsiterViewModel.Brand = user.Brand;
@@ -115,36 +112,52 @@ namespace ArvinTav.Areas.User.Controllers
                                 else
                                 {
 
-                                    if (user.Image != null)
+                                    var img = Bitmap.FromStream(ImageUp.InputStream);
+
+                                    if (img.Width > 600 || img.Height > 600)
                                     {
-                                        System.IO.File.Delete(Server.MapPath("/Document/img/User/" + user.Image));
+                                        ModelState.AddModelError("Image", "طول و عرض تصویر نباید بیشتر از 600 پیکسل باشد");
+                                        FullRegsiterViewModel fullReplyRegsiterViewModel = new FullRegsiterViewModel();
+                                        fullReplyRegsiterViewModel.FullName = user.FullName;
+                                        fullReplyRegsiterViewModel.Brand = user.Brand;
+                                        fullReplyRegsiterViewModel.Email = user.Email;
+                                        fullReplyRegsiterViewModel.Image = user.Image;
+                                        return View(fullRegsiterViewModel);
                                     }
+                                    else
+                                    {
 
-                                    UpdateUser.Image = Guid.NewGuid() + Path.GetExtension(ImageUp.FileName);
-                                    ImageUp.SaveAs(Server.MapPath("/Document/img/User/" + UpdateUser.Image));
+                                        if (user.Image != null)
+                                        {
+                                            System.IO.File.Delete(Server.MapPath("/Document/img/User/" + user.Image));
+                                        }
 
-                                    UpdateUser.PhoneNumber = User.Identity.Name;
-                                    UpdateUser.FullName = fullRegsiterViewModel.FullName;
-                                    UpdateUser.Email = fullRegsiterViewModel.Email;
-                                    UpdateUser.Brand = fullRegsiterViewModel.Brand;
-                                    userRepository.FullRegister(UpdateUser);
+                                        UpdateUser.Image = Guid.NewGuid() + Path.GetExtension(ImageUp.FileName);
+                                        ImageUp.SaveAs(Server.MapPath("/Document/img/User/" + UpdateUser.Image));
 
-                                    return RedirectToAction("Index");
+                                        UpdateUser.PhoneNumber = User.Identity.Name;
+                                        UpdateUser.FullName = fullRegsiterViewModel.FullName;
+                                        UpdateUser.Email = fullRegsiterViewModel.Email;
+                                        UpdateUser.Brand = fullRegsiterViewModel.Brand;
+                                        userRepository.FullRegister(UpdateUser);
+
+                                        return RedirectToAction("Index");
+
+                                    }
 
                                 }
 
                             }
+                            else
+                            {
+                                UpdateUser.PhoneNumber = User.Identity.Name;
+                                UpdateUser.FullName = fullRegsiterViewModel.FullName;
+                                UpdateUser.Email = fullRegsiterViewModel.Email;
+                                UpdateUser.Brand = fullRegsiterViewModel.Brand;
+                                userRepository.FullRegister(UpdateUser);
 
-                        }
-                        else
-                        {
-                            UpdateUser.PhoneNumber = User.Identity.Name;
-                            UpdateUser.FullName = fullRegsiterViewModel.FullName;
-                            UpdateUser.Email = fullRegsiterViewModel.Email;
-                            UpdateUser.Brand = fullRegsiterViewModel.Brand;
-                            userRepository.FullRegister(UpdateUser);
-
-                            return RedirectToAction("Index");
+                                return RedirectToAction("Index");
+                            }
                         }
                     }
                 }
